Move tests into appropriate test class
