+++ java/compaction/compaction-job-execution/src/test/java/sleeper/compaction/job/execution/CompactionTaskTest.java
@@ -21,21 +21,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import sleeper.compaction.job.CompactionJob;
-import sleeper.compaction.task.CompactionTaskFinishedStatus;
-import sleeper.compaction.task.CompactionTaskStatus;
-import sleeper.core.record.process.RecordsProcessed;
-import sleeper.core.record.process.RecordsProcessedSummary;
-
-import java.time.Instant;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Queue;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.compaction.job.CompactionJobStatusTestData.finishedCompactionRun;
-import static sleeper.compaction.job.CompactionJobStatusTestData.jobCreated;
-import static sleeper.compaction.job.CompactionJobStatusTestData.startedCompactionRun;
 
 public class CompactionTaskTest extends CompactionTaskTestBase {
 
@@ -84,134 +71,6 @@ public class CompactionTaskTest extends CompactionTaskTestBase {
             assertThat(successfulJobs).containsExactly(job1);
             assertThat(failedJobs).containsExactly(job2);
             assertThat(jobsOnQueue).isEmpty();
-        }
-    }
-
-    @Nested
-    @DisplayName("Update status stores")
-    class UpdateStatusStores {
-        @Test
-        void shouldSaveTaskAndJobWhenOneJobSucceeds() throws Exception {
-            // Given
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:01Z"), // Job started
-                    Instant.parse("2024-02-22T13:50:02Z"), // Job completed
-                    Instant.parse("2024-02-22T13:50:05Z"))); // Finish
-            CompactionJob job = createJobOnQueue("job1");
-
-            // When
-            RecordsProcessed recordsProcessed = new RecordsProcessed(10L, 10L);
-            runTask("test-task-1", processJobs(
-                    jobSucceeds(recordsProcessed)),
-                    times::poll);
-
-            // Then
-            RecordsProcessedSummary jobSummary = new RecordsProcessedSummary(recordsProcessed,
-                    Instant.parse("2024-02-22T13:50:01Z"),
-                    Instant.parse("2024-02-22T13:50:02Z"));
-            assertThat(taskStore.getAllTasks()).containsExactly(
-                    finishedCompactionTask("test-task-1",
-                            Instant.parse("2024-02-22T13:50:00Z"),
-                            Instant.parse("2024-02-22T13:50:05Z"),
-                            jobSummary));
-            assertThat(jobStore.getAllJobs(DEFAULT_TABLE_ID)).containsExactly(
-                    jobCreated(job, DEFAULT_CREATED_TIME,
-                            finishedCompactionRun("test-task-1", jobSummary)));
-        }
-
-        @Test
-        void shouldSaveTaskAndJobsWhenMultipleJobsSucceed() throws Exception {
-            // Given
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:01Z"), // Job 1 started
-                    Instant.parse("2024-02-22T13:50:02Z"), // Job 1 completed
-                    Instant.parse("2024-02-22T13:50:03Z"), // Job 2 started
-                    Instant.parse("2024-02-22T13:50:04Z"), // Job 2 completed
-                    Instant.parse("2024-02-22T13:50:05Z"))); // Finish
-            CompactionJob job1 = createJobOnQueue("job1");
-            CompactionJob job2 = createJobOnQueue("job2");
-
-            // When
-            RecordsProcessed job1RecordsProcessed = new RecordsProcessed(10L, 10L);
-            RecordsProcessed job2RecordsProcessed = new RecordsProcessed(5L, 5L);
-            runTask("test-task-1", processJobs(
-                    jobSucceeds(job1RecordsProcessed),
-                    jobSucceeds(job2RecordsProcessed)),
-                    times::poll);
-
-            // Then
-            RecordsProcessedSummary job1Summary = new RecordsProcessedSummary(job1RecordsProcessed,
-                    Instant.parse("2024-02-22T13:50:01Z"),
-                    Instant.parse("2024-02-22T13:50:02Z"));
-            RecordsProcessedSummary job2Summary = new RecordsProcessedSummary(job2RecordsProcessed,
-                    Instant.parse("2024-02-22T13:50:03Z"),
-                    Instant.parse("2024-02-22T13:50:04Z"));
-            assertThat(taskStore.getAllTasks()).containsExactly(
-                    finishedCompactionTask("test-task-1",
-                            Instant.parse("2024-02-22T13:50:00Z"),
-                            Instant.parse("2024-02-22T13:50:05Z"),
-                            job1Summary, job2Summary));
-            assertThat(jobStore.getAllJobs(DEFAULT_TABLE_ID)).containsExactlyInAnyOrder(
-                    jobCreated(job1, DEFAULT_CREATED_TIME,
-                            finishedCompactionRun("test-task-1", job1Summary)),
-                    jobCreated(job2, DEFAULT_CREATED_TIME,
-                            finishedCompactionRun("test-task-1", job2Summary)));
-        }
-
-        @Test
-        void shouldSaveTaskAndJobWhenOneJobFails() throws Exception {
-            // Given
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:01Z"), // Job started
-                    Instant.parse("2024-02-22T13:50:05Z"))); // Finish
-            CompactionJob job = createJobOnQueue("job1");
-
-            // When
-            runTask("test-task-1", processJobs(jobFails()), times::poll);
-
-            // Then
-            assertThat(taskStore.getAllTasks()).containsExactly(
-                    finishedCompactionTask("test-task-1",
-                            Instant.parse("2024-02-22T13:50:00Z"),
-                            Instant.parse("2024-02-22T13:50:05Z")));
-            assertThat(jobStore.getAllJobs(DEFAULT_TABLE_ID)).containsExactly(
-                    jobCreated(job, DEFAULT_CREATED_TIME,
-                            startedCompactionRun("test-task-1", Instant.parse("2024-02-22T13:50:01Z"))));
-        }
-
-        @Test
-        void shouldSaveTaskWhenNoJobsFound() throws Exception {
-            // Given
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:05Z"))); // Finish
-
-            // When
-            runTask("test-task-1", processNoJobs(), times::poll);
-
-            // Then
-            assertThat(taskStore.getAllTasks()).containsExactly(
-                    finishedCompactionTask("test-task-1",
-                            Instant.parse("2024-02-22T13:50:00Z"),
-                            Instant.parse("2024-02-22T13:50:05Z")));
-            assertThat(jobStore.getAllJobs(DEFAULT_TABLE_ID)).isEmpty();
-        }
-
-        private CompactionTaskFinishedStatus.Builder withJobSummaries(RecordsProcessedSummary... summaries) {
-            CompactionTaskFinishedStatus.Builder taskFinishedBuilder = CompactionTaskFinishedStatus.builder();
-            Stream.of(summaries).forEach(taskFinishedBuilder::addJobSummary);
-            return taskFinishedBuilder;
-        }
-
-        private CompactionTaskStatus finishedCompactionTask(String taskId, Instant startTime, Instant finishTime, RecordsProcessedSummary... summaries) {
-            return CompactionTaskStatus.builder()
-                    .startTime(startTime)
-                    .taskId(taskId)
-                    .finished(finishTime, withJobSummaries(summaries))
-                    .build();
         }
     }
 }
